# add the Hackaday Supercon ECP5 badge