Add the [Hackaday Supercon 2019 badge](https://hackaday.io/project/167255-2019-hackaday-superconference-badge) which has an ECP5 FPGA.

These changes are from [a fork](https://github.com/mwelling/litex-boards) by Michael Welling (@mwelling)

During Supercon, we tried two approaches:
- use the built-in 16MB QSPI SRAM
- use add-on cartiridge with 32MB SDRAM by Jacob Creedon

We were not able to get the QSPI SRAM working so I've removed those changes, and I have just added the changes that are needed to boot Linux with the 32MB SDRAM.

In addition to @mwelling, thank you to Jacob Creedon (@jcreedon), @GregDavill, Tim Ansell (@mithro), and Sean Cross (@xobs) who all helped get Linux working on this badge.

KiCad design files by @jcreedon for the SDRAM cartridge are [available on GitHub](https://github.com/jcreedon/dram-cart/).

There is also a [shared project](https://oshpark.com/shared_projects/IQSl2lid) to order the SDRAM cartridge PCB.

Refer to [my blog post](https://blog.oshpark.com/2019/12/20/boot-linux-on-this-hackaday-supercon-badge-with-this-sdram-cartridge/) for more information.

The Hackaday Supercon badge schematic and PCB layout are on [GitHub too](https://github.com/Spritetm/hadbadge2019_pcb).