@pdp7 2020-01-06
related PRs:
* https://github.com/litex-hub/linux-on-litex-vexriscv/pull/68
* https://github.com/enjoy-digital/litedram/pull/97