@enjoy-digital 2020-01-07
Thanks, i hope you all had fun working on that! It's merged with additional cleanup/simplifications:
https://github.com/litex-hub/litex-boards/commit/85c4f76ebaebef04d493c4ccedca55669d01c4cb
https://github.com/litex-hub/litex-boards/commit/c0e4578bea7f3faa3849df895dfe0cb8d498437f
(sorry i need to make this to still be able to maintain things :))

Since i don't have the hardware, can you check that things are still working? (i'll merge the linux-on-litex-vexriscv PR just after writing this).