@pdp7 2020-01-07
@enjoy-digital thank you for merging and making the cleanups.  I have tested and the badge still boots Linux correctly with lxterm and I can login OK as root:
```
pdp7@x1:~/dev/upstream/linux-on-litex-vexriscv$ lxterm --images=images.json /dev/ttyUSB0 --speed=1e6 --no-crc
```
```
[    0.000000] Linux version 5.0.13 (florent@lab) (gcc version 8.3.0 (Buildroot 2019.08-rc2-00011-gad9efda578)) #1 Thu Sep 12 14:20:26 CEST 2019
```