@pdp7 2020-01-07
@enjoy-digital fyi - I recorded an asciinema in case it it useful for reference: https://asciinema.org/a/292299