@enjoy-digital 2020-01-08
Thanks for the asciinema  record, i'll keep the link. Booting with SDRAM SoCs seems a bit slow, i'll try to identify the bottlenecks and speed things a bit.