@mithro 2020-01-08
From what I can see there is something misconfigured somewhere. Everything (not just booting) seems a *lot* slower than the OrangeCrab.