@pdp7 2020-01-08
I agree that is surprisingly slow.  Can the DDR on the @gregdavill OrangeCrab really make that big of a difference?