@gregdavill 2020-01-09
The SDRAM is running with the following specifications:
 * 8bit data bus
 * 48MHz clock
 * Raw bus throughput 384 Mbit/s (8 * 48M)

The DDR3 on the OrangeCrab:
 * 16bit data bus
 * 96MHz DDR Clock. (2:1 clock gearing)
 * Raw bus throughput 3,072Mbit/s (16 * 96M * 2)

Of course this is ignoring latency and other differences between the DDR3 and the SDRAM. But there is a pretty big difference in the raw bus speed.