@pdp7 2020-01-09
@gregdavill thanks for the details specs.  That is good to know there is an order of magnitude difference in raw bus speed.