@xobs 2020-01-09
Side note, but having bursts on the Wishbone bus to avoid the /2 factor would be super awesome for lots of other projects.

On 9 January 2020 17:41:51 GMT+08:00, enjoy-digital <notifications@github.com> wrote:
>From the CPU side, the maximal throughput on the wishbone bus is :
>48M*32/2=768Mbit/s. (/2 since a wishbone access takes minimum 2
>cycles). I need to look at how the 32-bit are down-converted to 8-bit
>for the SDRAM accesses, it's probably not as efficient as it could be.
>(not sure i ever try to optimize this case since on most designs
>SDRAM's controller data width >= 32). A relatively easy next thing to
>do to speed things with SDRAM would be to run the SDRAM at 96MHz
>instead of 48MHz. Then the next optimization would be to support bursts
>on the wishbone bus to avoid the /2 factor.
>
>-- 
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/litex-hub/litex-boards/pull/31#issuecomment-572476595