@enjoy-digital 2020-01-09
@xobs: indeed, for now it was not that critical since the CPU was mostly used to assist the hardware and do some control, but now that we are running OSes on it's becoming limiting, i'll look at the options.