@enjoy-digital 2020-01-10
With https://github.com/enjoy-digital/litedram/commit/34e6c24d7296aeed6fb71ddc009b357cfc5a82ca and https://github.com/enjoy-digital/litex/commit/fa22d6aa82ee145671a424c7b3112daa82c9ca0d we have a ~10% boot time speedup on designs using SDRAM:
- De0Nano: 94.6.s to  84.6s
- ULX3S: 75.9s to 68.2s

On Arty with DDR3 the gain is effect more limited: 8.7s to 8.4s. That would be interesting to test this on the badge.